cache: Refactor out cacheRecord equal* fields
The equalMutable and equalImmutable fields were used to implement the
"lazy commit" feature (not to be confused with lazy refs), which allows
refs to put off comitting their snapshots to the underlying driver until
absolutely needed.

However, this way of modeling it resulted in there being 2 actual
cacheRecords even when the underlying snapshot was the same. This also
lead to a lot of convoluted logic checking for the existence of the
equal* fields.

This commits switches to a new approach of only maintaining a single
cacheRecord for what used to be separate equal* records. The cacheRecord
internally keeps track of whether it has been finalized or not and
updates which snapshot ID is underlying it.

This results in overall much simpler logic, as can be seen by looking at
the before+after of GetMutable, prune, and release, among others. The
only 2 meaningfully-sized blocks of code added are
1. In cacheManager.init, to handle migrating records from the old format
2. In cacheManager.getRecord, to handle a corner-case involving the
   immutableRef.finalize method

The first of those can eventually be deleted once the older versions of
buildkit are no longer supported and the second is a very minor
tradeoff given the rest of simplifications elsewhere.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>