cache: refactor cache record mount cache
This is mostly just preparation for merge-op. The lowercase "extract"
method is replaced with a more general "prepareMounts" method which does
all the work necessary to make a cache record mountable and then caches
the mountable object, ready to use whenever Mounts is called in the
future. It centralizes more of the logic around setting up these mounts,
which will make the merge-op handling code a bit more tractable.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>