cache: Replace (Im)MutableRef interfaces w/ structs.
Previously, cache.ImmutableRef and cache.MutableRef were interfaces
returned by the cache package, but this was not really useful as many of
the methods in cache.Manager would turn them into their internal structs
when passed as args (either through type coercion or by looking up the
internal struct based on the interface object's ID).

The code for converting the interface object to the internal struct was
confusing, defeated the purpose of interfaces and resulted in wasted
time checking entire ref chains for lazy blobs as part of the Get
method (which may soon get even worse w/ merge and diff refs).

Now, those interfaces are just plain structs, allowing their internal
fields+methods to be directly accessed when passed to the cache manager.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>