Migrate some logic from cache to snapshot package.
Buildkit's Snapshotter interface is now updated to handle more snapshot
management, allowing more snapshot-specific logic to be migrated from
the cache package to the snapshot package.

The biggest change is that outside callers only need to manage 1 key
for an underlying snapshot rather than 3 (1 for each of the active,
commit and view of a snapshot). The snapshotter now internally manages
switching to different keys for each of the possible states based on the
single key provided by callers. This additionally allows the snapshotter
to hide the existence of Views entirely and instead just internally
create Views when the caller requests the mounts of a committed
snapshot.

The snapshotter now also provides methods for managing the leases of
snapshots.

These changes simplify the cache code as is but will become even more
useful for the merge-op changes, which will introduce significant further
complication to snapshot management.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>