Fix rootless merge on pre 5.11 patched kernels.
Kernels that enabled rootless overlay mounts before official support in
v5.11 via patches have behavior such that opaque xattrs are created on
dirs but they aren't visible inside the user namespace. This breaks
checks for unnecessary opaques. This change results in the efficient
overlay merge being disabled in this specific case, which avoids the
issue for those kernels at the cost of performance.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>