snapshot: Change Merge to accept Diffs.
Previously, the snapshot Merge method accepted a list of parent snapshot
IDs to merge together. This commit updates it to instead accept Diff
structs, which allows the caller to specify precisely which diffs to
apply. For the basic merge-op case, each Diff.Upper will be set to the
snapshot ID of the merge input and Diff.Lower will be empty string. The
logic of snapshot Merge has been updated to expand such diffs in such a
way that it has the exact same behavior as before. So, this change
results in no new behavior yet, but it enables us to add in DiffOp
support on top of this more generalized Merge method in a subsequent
commit.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>