cache: add support for Diff refs.
This allows you to create refs that are single layers representing the
diff between any two arbitrary refs. The primary use case for this is
to allows users to extract the changes created by ops like Exec and
rebase them elsewhere through MergeOp. However, there is no restriction
on the inputs to DiffOp and the resulting ref's layer is simply the
layer created by running the differ on the two inputs refs
(specifically, the same differ used during exports).

A Diff ref can be mounted by itself, in which case it is defined as the
result of applying the diff to Scratch. Most use cases though will use
Diff refs as the input to a MergeOp, in which case the diff is just
applied on top of the lower merge inputs, as was the case before.

In cases like Diff(A, A->B) (i.e. cases where the diff is between two
refs where the upper is only one layer different than the lower), the
blob representing the diff will be re-used from the upper-most layer, as
it is always exactly equivalent.

However, cases such as Diff(A, A->B->C) don't re-use input blobs. It
would have been possible to define such cases as B->C, i.e. diff would
extract out the "slice" of multiple layers that separate the lower and
upper refs. Doing so would change the definition of what Diff is though,
likely resulting in an interface that's more confusing and limiting for
end-users.

This is because there's a distinction between running the differ between
two refs and merging the layers the separate them. Namely, there can be
deletes that exist in the layers between the two refs that don't exist in
the overall diff between them. Such deletes can make a difference once
rebased ontop of other merge inputs. Additionally, in some use cases the
size of the layers separating two refs may be larger than a single blob
representing the overall diff (specificially in cases where files are
overwritten across many layers). So it's not always desireable to re-use
every possible blob in the diff between the refs.

Instead, we keep the definition of Diff refs as simple as possible: just
the result of running the differ between the lower and upper input refs.
This means that blobs will only be re-used in the single-layer diff
cases, all others will result in the diff ref generating its own blob.
Users that want something more like the "sliceable" diff as described
above can achieve that by just merging together diffs client-side, but
those who don't won't have to think about it.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>